requirements.txt
langchain==0.1.12
langchain-core==0.1.32
langchain-community==0.0.27
llama-cpp-python==0.2.56
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
requests==2.31.0
